ref(performance): Extract `<TraceView>` into a component
We are looking to use TraceView in Replays, so we extracted it into a new component (as well as some others that were being rendered via class methods)
